refactor: apply INIT_CWD for all bin scripts
