+++ bin/generate-android-snapshot
