+++ index.js
@@ -1,9 +1,9 @@
 const path = require("path");
 const { existsSync } = require("fs");
 
-const { getPackageJson, isAngular } = require("./projectHelpers");
+const { getPackageJson, getProjectDir, isAngular } = require("./projectHelpers");
 
-const PROJECT_DIR = path.dirname(path.dirname(__dirname));
+const PROJECT_DIR = getProjectDir({ nestingLvl: 2 });
 const APP_DIR = path.join(PROJECT_DIR, "app");
 
 Object.assign(exports, require('./plugins'));
